backend/requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0

# JWT Authentication
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1

# Database
mysqlclient==2.2.6

# Celery + Redis
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Email
django-templated-mail==1.1.1

# Utils
python-dotenv==1.0.1
django-filter==24.3
Pillow==11.0.0
gunicorn==23.0.0

# Testing
factory-boy==3.3.1
pytest-django==4.9.0
